+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2025-06-16
+
 ### Changed
 
 - Authentication using IAM service instead of Core
@@ -53,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TrustedSource model and TrustedSourceCollection
 - UserJob model and UserJobCollection
 
-[Unreleased]: https://github.com/cytomine/Cytomine-js-client/compare/3.0.1..HEAD
+[Unreleased]: https://github.com/cytomine/Cytomine-js-client/compare/3.1.0..HEAD
+[3.1.0]: https://github.com/cytomine/Cytomine-js-client/releases/tag/3.1.0
 [3.0.1]: https://github.com/cytomine/Cytomine-js-client/releases/tag/3.0.1
 [3.0.0]: https://github.com/cytomine/Cytomine-js-client/releases/tag/3.0.0